feat: qna controller 추가

+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/study/repository/StudyNoticeRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.ssafy.lighthouse.domain.study.entity.StudyNotice;
 
-public interface StudyNoticeRepository extends JpaRepository<StudyNotice, Integer> {
+public interface StudyNoticeRepository extends JpaRepository<StudyNotice, Long> {
 	@EntityGraph(attributePaths = {"studyNoticeChecks"})
 	@Query(value = "SELECT sn FROM StudyNotice sn WHERE  sn.isValid = 1 AND sn.studyId = :studyId")
 	public List<StudyNotice> findByStudyId(@Param("studyId") Long studyId);
